fix type error + fix makefile

+++ fft.h
@@ -25,8 +25,8 @@ double fft(
     int64_t input_1_R[],
     int64_t output_L[],
     int64_t output_R[],
-    double *dft_mem_L,
-    double *dft_mem_R,
+    float *dft_mem_L,
+    float *dft_mem_R,
     struct transfer_function *curve_processed,
     int level);
